refactor: 디렉토리 구조 변경
- 기능에 맞게 분리
