--- src/main/java/com/sedin/qna/common/LoginRequired.java
@@ -1,15 +1,0 @@
-package com.sedin.qna.common;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD, ElementType.ANNOTATION_TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface LoginRequired {
-
-}
-
