--- src/main/java/com/sedin/qna/config/WebMvcConfig.java
@@ -1,29 +1,0 @@
-package com.sedin.qna.config;
-
-import com.sedin.qna.interceptor.AuthenticationInterceptor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.util.ResourceUtils;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    private final AuthenticationInterceptor authenticationInterceptor;
-
-    public WebMvcConfig(AuthenticationInterceptor authenticationInterceptor) {
-        this.authenticationInterceptor = authenticationInterceptor;
-    }
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/api/docs/**")
-                .addResourceLocations(ResourceUtils.CLASSPATH_URL_PREFIX + "/static/docs/");
-    }
-
-    @Bean
-    public WebMvcConfigurer webMvcConfigurerInterceptors() {
-        return new WebConfigInterceptors(authenticationInterceptor);
-    }
-}
