--- src/main/java/com/sedin/qna/util/JwtUtil.java
@@ -1,46 +1,0 @@
-package com.sedin.qna.util;
-
-import com.sedin.qna.exception.InvalidTokenException;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import io.jsonwebtoken.security.SignatureException;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-
-@Component
-public class JwtUtil {
-
-    public final static String ACCOUNT_ID = "accountId";
-    private final Key key;
-
-    public JwtUtil(@Value("${jwt.secret}") String secret) {
-        key = Keys.hmacShaKeyFor(secret.getBytes());
-    }
-
-    public String encode(Long accountId) {
-        return Jwts.builder()
-                .claim(ACCOUNT_ID, accountId)
-                .signWith(key)
-                .compact();
-    }
-
-    public Claims decode(String token) {
-        if (token == null || StringUtils.isBlank(token)) {
-            throw new InvalidTokenException();
-        }
-
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-        } catch (SignatureException e) {
-            throw new InvalidTokenException();
-        }
-    }
-}
